pull from main

+++ deps/icu-small/source/tools/toolutil/package.cpp
@@ -1,3 +1,5 @@
+// © 2016 and later: Unicode, Inc. and others.
+// License & terms of use: http://www.unicode.org/copyright.html
 /*
 *******************************************************************************
 *
@@ -6,7 +8,7 @@
 *
 *******************************************************************************
 *   file name:  package.cpp
-*   encoding:   US-ASCII
+*   encoding:   UTF-8
 *   tab size:   8 (not used)
 *   indentation:4
 *
@@ -1288,11 +1290,11 @@ void Package::setItemCapacity(int32_t max)
   Item *oldItems = items;
   if(newItems == NULL) {
     fprintf(stderr, "icupkg: Out of memory trying to allocate %lu bytes for %d items\n",
-        (unsigned long)max*sizeof(items[0]), max);
+        (unsigned long)(max*sizeof(items[0])), max);
     exit(U_MEMORY_ALLOCATION_ERROR);
   }
   if(items && itemCount>0) {
-    uprv_memcpy(newItems, items, itemCount*sizeof(items[0]));
+    uprv_memcpy(newItems, items, (size_t)itemCount*sizeof(items[0]));
   }
   itemMax = max;
   items = newItems;
